requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==24.2.0
autoflake==2.3.1
black==25.1.0
blinker==1.9.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.8
coloredlogs==15.0.1
distlib==0.4.0
elastic-transport==8.17.0
elasticsearch==8.17.1
envparse==0.2.0
fake-useragent==2.0.2
filelock==3.18.0
frozenlist==1.5.0
gigachat==0.1.39.post1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
identify==2.6.12
idna==3.10
importlib_resources==6.4.5
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
octodiary==0.3.0
packaging==24.2
pathspec==0.12.1
phonenumbers==8.13.52
platformdirs==4.3.6
pre_commit==4.2.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.4.0
python-logstash==0.4.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
virtualenv==20.32.0
Werkzeug==3.1.3
yarl==1.17.2
